lib/receipt-processing/python-processor/requirements.txt
opencv-python==4.8.1.78
pytesseract==0.3.10
pyzbar==0.1.9
pillow==10.1.0
numpy==1.24.3
